Extract rollup-related utils
